requirements.txt
absl-py==1.0.0
allennlp==2.9.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
attrs==21.4.0
backcall==0.2.0
backports.csv==1.0.7
base58==2.1.1
beautifulsoup4==4.11.1
benepar==0.2.0
bleach==5.0.0
blis==0.7.5
boto3==1.20.51
botocore==1.23.51
cached-path==1.0.2
cachetools==5.0.0
catalogue==2.0.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
checklist==0.0.11
cheroot==8.6.0
CherryPy==18.6.1
click==7.1.2
comet-ml==3.24.0
configobj==5.0.6
cycler==0.11.0
cymem==2.0.6
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
demjson==1.6
dill==0.3.4
docker-pycreds==0.4.0
dulwich==0.20.30
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.2.0/en_core_web_sm-3.2.0-py3-none-any.whl
entrypoints==0.4
everett==2.0.1
executing==0.8.3
fairscale==0.4.5
fastjsonschema==2.15.3
feedparser==6.0.8
filelock==3.4.2
flatbuffers==2.0
fonttools==4.33.3
future==0.18.2
gast==0.5.3
gitdb==4.0.9
GitPython==3.1.26
google-api-core==2.5.0
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.2.2
google-cloud-storage==1.44.0
google-crc32c==1.3.0
google-pasta==0.2.0
google-resumable-media==2.2.0
googleapis-common-protos==1.54.0
grpcio==1.43.0
h5py==3.6.0
huggingface-hub==0.2.1
idna==3.3
iniconfig==1.1.1
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
iso-639==0.4.5
jaraco.classes==3.2.1
jaraco.collections==3.5.1
jaraco.context==4.1.1
jaraco.functools==3.5.0
jaraco.text==3.7.0
jedi==0.18.1
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.1.0
jsonnet==0.18.0
jsonschema==4.5.1
jupyter==1.0.0
jupyter-client==7.3.1
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
langcodes==3.3.0
libclang==13.0.0
lmdb==1.3.0
lxml==4.8.0
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
mkl-fft==1.3.1
mkl-random==1.22
mkl-service==2.4.0
more-itertools==8.13.0
munch==2.5.0
murmurhash==1.0.6
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
nltk==3.6.5
notebook==6.4.11
numpy==1.21.2
nvidia-ml-py3==7.352.0
oauthlib==3.2.0
olefile==0.46
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pathy==0.6.1
patternfork-nosql==3.6
pdfminer.six==20211012
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pluggy==1.0.0
portend==3.1.0
preshed==3.0.6
prometheus-client==0.14.1
promise==2.3
prompt-toolkit==3.0.29
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.8.2
Pygments==2.12.0
pyparsing==3.0.8
pyrsistent==0.18.1
pytest==7.1.2
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2022.1
PyYAML==6.0
pyzmq==23.0.0
qtconsole==5.3.0
QtPy==2.1.0
regex==2022.3.15
requests==2.27.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rsa==4.8
s3transfer==0.5.1
sacremoses==0.0.47
scikit-learn==1.1.1
scipy==1.8.1
semantic-version==2.8.5
Send2Trash==1.8.0
sentencepiece==0.1.96
sentry-sdk==1.5.4
sgmllib3k==1.0.0
shortuuid==1.0.8
six==1.16.0
smart-open==5.2.1
smmap==5.0.0
soupsieve==2.3.2.post1
spacy==3.2.1
spacy-legacy==3.0.8
spacy-loggers==1.0.1
srsly==2.4.2
stack-data==0.2.0
tempora==5.0.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==1.2
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
terminado==0.15.0
tf-estimator-nightly==2.8.0.dev2021122109
thinc==8.0.13
threadpoolctl==3.1.0
tinycss2==1.1.1
tokenizers==0.10.3
tomli==2.0.1
torch==1.10.2+cu113
torch-struct==0.5
torchaudio==0.10.2+cu113
torchtext==0.11.2
torchvision==0.11.3+cu113
tornado==6.1
tqdm==4.64.0
traitlets==5.2.1.post0
transformers==4.15.0
typer==0.4.0
typing-extensions==3.10.0.2
urllib3==1.26.9
wandb==0.12.10
wasabi==0.9.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
widgetsnbextension==3.6.0
wrapt==1.14.1
wurlitzer==3.0.2
yaspin==2.1.0
zc.lockfile==2.0
